pyproject.toml
[project]
name = "aura"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.120.0",
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.46.0",
    "piper-tts>=1.3.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
